# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend for the all rated extreme demon list",
        "title": "Aredl API",
        "contact": {
            "name": "Aredl",
            "url": "https://aredl.net/"
        },
        "version": "1.0"
    },
    "host": "api.aredl.net",
    "basePath": "/api",
    "paths": {
...
        "/aredl/names": {
            "get": {
                "description": "Gives a map of important users grouped by their role. This also includes aredl plus members",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "Important users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/aredl.NameUser"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/unban": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Unbans a user\nRequires user permission: user_ban\nAdditionally the user needs to be able to affect the user with their permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Unban user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "internal user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "aredl.NameUser": {
            "type": "object",
            "properties": {
                "global_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "util.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Perform actions as a user. It is also used to access endpoints that require user permissions.",
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    }
}